speed up Subcase deepcopy
Control decks with a lot of subcases are slow to read, due to deepcopying of previously defined subcases.  A __deepcopy__ method has been added to the Subcase class to override the default copy.deepcopy method.  This method is at least 2x faster, although ideally it would be an order of magnitude faster.