default subcase is now shallow copied
When the default subcase is copied for every subcase that is defined, all SET definitions were previously deep copied.  If these sets were large, this would be very slow.  Now the default subcase is shallow copied, and the bdf read time is greatly improved.  A unit test will be added to confirm that it is working correctly.