fixes #493 - CBUSH cid default
The code was previously setting cid to 0 if blank when x1_g0 is a float.  This will always override the cgap x-axis if the user wishes to use the ga to gb vector for its x-axis where ga and gb are not coincident (in which case cid would be blank).  This causes incorrect results.