bdf reader: StringIO parse headers
The bdf reader doesn't parse headers for StringIO buffers passed in the 'read_bdf' method.  Small changes were made to parse the headers the StringIO objects.  After the 'readlines' method is called on the StringIO object, it is rewound using seek(0) to go back to the beginning so that the lines can be read again in the '_get_main_lines' method.